fix: NDArray.cs - all params int[] methods now have long[] primary + int[] convenience
Pattern applied:
- Get*(params long[] indices) - primary implementation calling Storage
- Get*(params int[] indices) - delegates to long[] via Shape.ComputeLongShape()
- Set*(value, params long[] indices) - primary implementation
- Set*(value, params int[] indices) - delegates to long[] version

Covers: GetData, GetBoolean, GetByte, GetChar, GetDecimal, GetDouble,
GetInt16, GetInt32, GetInt64, GetSingle, GetUInt16, GetUInt32, GetUInt64,
GetValue, GetValue<T>, SetData (3 overloads), SetValue (3 overloads),
SetBoolean, SetByte, SetInt16, SetUInt16, SetInt32, SetUInt32, SetInt64,
SetUInt64, SetChar, SetDouble, SetSingle, SetDecimal

Related to #584

+++ src/NumSharp.Core/Backends/Iterators/NDIterator.cs
@@ -210,7 +210,7 @@ namespace NumSharp
                     {
                         var hasNext = new Reference<bool>(true);
                         var iterator = new ValueCoordinatesIncrementor(ref shape, delegate(ref ValueCoordinatesIncrementor _) { hasNext.Value = false; });
-                        Func<int[], int> getOffset = shape.GetOffset;
+                        Func<long[], long> getOffset = shape.GetOffset;
                         var index = iterator.Index;
 
                         MoveNext = () =>
@@ -333,7 +333,7 @@ namespace NumSharp
                     {
                         var iterator = new ValueCoordinatesIncrementor(ref shape, delegate(ref ValueCoordinatesIncrementor incr) { incr.Reset(); });
                         var index = iterator.Index;
-                        Func<int[], int> getOffset = shape.GetOffset;
+                        Func<long[], long> getOffset = shape.GetOffset;
                         MoveNext = () =>
                         {
                             var ret = *((TOut*)localBlock.Address + getOffset(index));
